Fixing missing status translations for groups, users & categories.
For the edit.blade and list.blade files, I think this is more of a bandaid than an actual fix. My theory is something is sending the "inactive" status improperly, so PM4 is interpreting it everywhere as "inactive" as opposed to "Inactive", and that capital "I" is required for the translation to work correctly.

JIRA: https://processmaker.atlassian.net/browse/FOUR-276