FOUR-22703
- Create an user with manager in USER configuration
- Create a process
- Add Start Event → Task → End Event
- Configure the task form with "Assignee manager escalation"
- Create a cases
- Route to the task open the task
- Click on ESCALATE TO MANAGER button

It is not possible to Escalate to manager with "Assignee Manager Escalation"

The way the script elements are loaded has changed, so the corresponding missing definition is added.

https://processmaker.atlassian.net/browse/FOUR-22703
FOUR-22703

- Create an user with manager in USER configuration
- Create a process
- Add Start Event → Task → End Event
- Configure the task form with "Assignee manager escalation"
- Create a cases
- Route to the task open the task
- Click on ESCALATE TO MANAGER button

It is not possible to Escalate to manager with "Assignee Manager Escalation"

The way the script elements are loaded has changed, so the corresponding missing definition is added.

https://processmaker.atlassian.net/browse/FOUR-22703