FOUR-23734 [44349] - Reassignment User List Not Filtering by Regular Expressio
 ## Description:
Steps to Reproduce the Issue:
  Open a Request/Case:
Navigate to the ProcessMaker cases and open an existing request (case).
Click on the Reassignment Button:
Locate and click the "Reassignment" button to reassign the case to another user.
  View the Assignment User List:
Observe the list of users displayed in the reassignment dropdown.
  Current Behavior (Issue):
Instead of filtering users as per the regular expression, all users are displayed in the reassignment list.
  Expected Behavior:
The user list should be filtered based on the regular expression condition set in the process.
Only users that match the specified regular expression should appear in the list.
  Impact:
The client expects the user list to be filtered dynamically based on the regular expression condition applied in the process, but it is not working correctly.
  Potentials Fix:
Request the development team to investigate and ensure that the reassignment user list correctly applies the regular expression condition as configured in the process.
  Acceptance criteria:
When having a Rule Expression, the reassignment should only allow to reassign to Admin, Process Manager or the users listed in the Rule Expression list.

 ## Related Tickets & Packages
https://processmaker.atlassian.net/browse/FOUR-23734