FOUR-24591 [45224]Script Execution fails when the environment variables reach the limit of data size  ## Description: - Steps to Reproduce: Go-to Environmental Variables. Add a new environmental variable. In the value field ensure the character length is equal to or above 5000 or more (you can use a character generator). Save it Run a simple PHP  “Hello World“ Script.
- Current Behavior:
When you run the script, you get an error below ( I have attached a file, due to the length of the error)

- Expected Behavior:
There should not be any limit on adding environment variables based on the size, or if there is, it must be part of the documentation, or if this is related to any other parameter to be increased must also be documented.
On the other hand, this should not affect the script's execution, or at least there should be a message that the number or size has been exceeded.

- Notes to QA:
The error was discovered when initially testing the addition of a PGP key as an environmental variable.
Further investigations made us detect that adding new values to an environmental variable of large length lead to error being returned.
The use case of our client was having more than 30 PGP keys stored.
Not using microservices.

 ## Related Tickets & Packages
https://processmaker.atlassian.net/browse/FOUR-24591